[PodsProjectGenerator] Group support files generation logic in SupportFilesGenerator
