+++ lib/cocoapods/installer/pods_project_generator/target_installer.rb
@@ -59,34 +59,14 @@ module Pod
           target.target = @native_target
         end
 
-        # Creates the group that holds the references to the support files
-        # generated by this installer.
-        #
-        # @return [void]
-        #
-        def create_suport_files_group
-          @support_files_group = project.support_files_group.new_group(target.name)
-        end
 
-        # Generates a dummy source file for each target so libraries that contain
-        # only categories build.
-        #
-        # @return [void]
-        #
-        def create_dummy_source
-          UI.message "- Generating dummy source file" do
-            path = target.dummy_source_path
-            generator = Generator::DummySource.new(target.label)
-            generator.save_as(path)
-            file_reference = add_file_to_support_group(path)
-            target.target.source_build_phase.add_file_reference(file_reference)
-          end
-        end
 
         # @return [PBXNativeTarget] the target generated by the installation
         #         process.
         #
         # @note   Generated by the {#add_target} step.
+        #
+        # TODO Remove
         #
         attr_reader :native_target
 
@@ -100,28 +80,6 @@ module Pod
         #
         def project
           sandbox.project
-        end
-
-        # @return [TargetDefinition] the target definition of the target.
-        #
-        def target_definition
-          target.target_definition
-        end
-
-        # @return [PBXGroup] the group where the file references to the support
-        #         files should be stored.
-        #
-        attr_reader :support_files_group
-
-        # Adds a reference to the given file in the support group of this target.
-        #
-        # @param  [Pathname] path
-        #         The path of the file to which the reference should be added.
-        #
-        # @return [PBXFileReference] the file reference of the added file.
-        #
-        def add_file_to_support_group(path)
-          support_files_group.new_file(path)
         end
 
         #-----------------------------------------------------------------------#
